+++ cmp/form/FormModel.ts
@@ -245,7 +245,6 @@ export class FormModel extends HoistModel {
         const states = map(this.fields, m => m.validationState);
         if (states.includes('NotValid')) return 'NotValid';
         if (states.includes('Unknown')) return 'Unknown';
-        if (states.includes('ValidWithWarnings')) return 'ValidWithWarnings';
         return 'Valid';
     }
 
@@ -257,7 +256,7 @@ export class FormModel extends HoistModel {
 
     /** True if all fields are valid. */
     get isValid(): boolean {
-        return this.validationState === 'Valid' || this.validationState === 'ValidWithWarnings';
+        return this.validationState === 'Valid';
     }
 
     /** List of all validation errors for this form. */