+++ cmp/input/HoistInputModel.ts
@@ -6,10 +6,11 @@
  */
 import {FieldModel} from '@xh/hoist/cmp/form';
 import {DefaultHoistProps, HoistModel, HoistModelClass, useLocalModel} from '@xh/hoist/core';
+import {Validation, ValidationSeverity} from '@xh/hoist/data';
 import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
 import {createObservableRef} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
-import {isEqual} from 'lodash';
+import {groupBy, isEmpty, isEqual} from 'lodash';
 import {FocusEvent, ForwardedRef, ReactElement, ReactInstance, useImperativeHandle} from 'react';
 import {findDOMNode} from 'react-dom';
 import './HoistInput.scss';
@@ -337,8 +338,18 @@ export function useHoistInputModel(
 
     const field = inputModel.getField(),
         validityClass = field?.isNotValid && field?.validationDisplayed ? 'xh-input-invalid' : null,
+        validationsBySeverity = groupBy(field?.validations, 'severity') as Record<
+            ValidationSeverity,
+            Validation[]
+        >,
         warningClass =
-            field?.isValidWithWarnings && field?.validationDisplayed ? 'xh-input-warning' : null,
+            !isEmpty(validationsBySeverity.warning) && field?.validationDisplayed
+                ? 'xh-input-warning'
+                : null,
+        infoClass =
+            !isEmpty(validationsBySeverity.info) && field?.validationDisplayed
+                ? 'xh-input-info'
+                : null,
         disabledClass = props.disabled ? 'xh-input-disabled' : null;
 
     return component({
@@ -347,8 +358,7 @@ export function useHoistInputModel(
         ref: inputModel.domRef,
         className: classNames(
             'xh-input',
-            validityClass,
-            warningClass,
+            validityClass ?? warningClass ?? infoClass,
             disabledClass,
             props.className
         )