+++ data/impl/StoreValidator.ts
@@ -9,7 +9,11 @@ import {HoistBase} from '@xh/hoist/core';
 import {computed, makeObservable, runInAction, observable} from '@xh/hoist/mobx';
 import {sumBy, chunk} from 'lodash';
 import {findIn} from '@xh/hoist/utils/js';
-import {RecordValidationMessagesMap, RecordValidator} from './RecordValidator';
+import {
+    RecordValidationMessagesMap,
+    RecordValidationsMap,
+    RecordValidator
+} from './RecordValidator';
 import {ValidationState} from '../validation/ValidationState';
 import {Store} from '../Store';
 import {StoreRecordId} from '../StoreRecord';
@@ -21,16 +25,10 @@ import {StoreRecordId} from '../StoreRecord';
 export class StoreValidator extends HoistBase {
     store: Store;
 
-    /** True if the store is confirmed to be Valid (with or without warnings). */
+    /** True if the store is confirmed to be Valid. */
     @computed
     get isValid(): boolean {
-        return this.validationState === 'Valid' || this.validationState === 'ValidWithWarnings';
-    }
-
-    /** True if the store is confirmed to be Valid but has warnings. */
-    @computed
-    get isValidWithWarnings(): boolean {
-        return this.validationState === 'ValidWithWarnings';
+        return this.validationState === 'Valid';
     }
 
     /** True if the store is confirmed to be NotValid. */
@@ -57,16 +55,10 @@ export class StoreValidator extends HoistBase {
         return sumBy(this.validators, 'errorCount');
     }
 
-    /** Map of StoreRecord IDs to StoreRecord-level warning maps. */
+    /** Map of StoreRecord IDs to StoreRecord-level validations maps. */
     @computed.struct
-    get warnings(): StoreValidationMessagesMap {
-        return this.getWarningMap();
-    }
-
-    /** Count of all validation warnings for the store. */
-    @computed
-    get warningCount(): number {
-        return sumBy(this.validators, 'warningCount');
+    get validations(): StoreValidationsMap {
+        return this.getValidationsMap();
     }
 
     /** True if any records are currently recomputing their validation state. */
@@ -106,7 +98,6 @@ export class StoreValidator extends HoistBase {
         const states = this.mapValidators(v => v.validationState);
         if (states.includes('NotValid')) return 'NotValid';
         if (states.includes('Unknown')) return 'Unknown';
-        if (states.includes('ValidWithWarnings')) return 'ValidWithWarnings';
         return 'Valid';
     }
 
@@ -117,9 +108,9 @@ export class StoreValidator extends HoistBase {
         return ret;
     }
 
-    getWarningMap(): StoreValidationMessagesMap {
-        const ret: StoreValidationMessagesMap = {};
-        this._validators.forEach(v => (ret[v.id] = v.warnings));
+    getValidationsMap(): StoreValidationsMap {
+        const ret: StoreValidationsMap = {};
+        this._validators.forEach(v => (ret[v.id] = v.validations));
         return ret;
     }
 
@@ -181,5 +172,7 @@ export class StoreValidator extends HoistBase {
     }
 }
 
-/** Map of StoreRecord IDs to StoreRecord-level error maps. */
+/** Map of StoreRecord IDs to StoreRecord-level messages maps. */
 export type StoreValidationMessagesMap = Record<StoreRecordId, RecordValidationMessagesMap>;
+/** Map of StoreRecord IDs to StoreRecord-level validations maps. */
+export type StoreValidationsMap = Record<StoreRecordId, RecordValidationsMap>;